Move Jupyter Hub Code into a separate cls
