+++ src/standalone/userJupyterServer/jupyterPasswordConnect.ts
@@ -3,15 +3,10 @@
 
 import { CancellationError, ConfigurationTarget, QuickInputButtons } from 'vscode';
 import { IApplicationShell } from '../../platform/common/application/types';
-import {
-    IAsyncDisposableRegistry,
-    IConfigurationService,
-    IDisposable,
-    IDisposableRegistry
-} from '../../platform/common/types';
+import { IConfigurationService, IDisposable, IDisposableRegistry } from '../../platform/common/types';
 import { DataScience } from '../../platform/common/utils/localize';
 import { noop } from '../../platform/common/utils/misc';
-import { IMultiStepInputFactory, IMultiStepInput, InputFlowAction } from '../../platform/common/utils/multiStepInput';
+import { InputFlowAction } from '../../platform/common/utils/multiStepInput';
 import { traceWarning } from '../../platform/logging';
 import { sendTelemetryEvent, Telemetry } from '../../telemetry';
 import {
@@ -20,7 +15,6 @@ import {
     IJupyterServerUriEntry,
     IJupyterServerUriStorage
 } from '../../kernels/jupyter/types';
-import { Deferred, createDeferred } from '../../platform/common/utils/async';
 import { disposeAllDisposables } from '../../platform/common/helpers';
 
 export interface IJupyterPasswordConnectInfo {
@@ -38,8 +32,6 @@ export class JupyterPasswordConnect {
     constructor(
         private appShell: IApplicationShell,
 
-        private readonly multiStepFactory: IMultiStepInputFactory,
-        private readonly asyncDisposableRegistry: IAsyncDisposableRegistry,
         private readonly configService: IConfigurationService,
         private readonly agentCreator: IJupyterRequestAgentCreator | undefined,
         private readonly requestCreator: IJupyterRequestCreator,
@@ -49,10 +41,6 @@ export class JupyterPasswordConnect {
         // Sign up to see if servers are removed from our uri storage list
         this.serverUriStorage.onDidRemove(this.onDidRemoveUris, this, this.disposables);
     }
-    private static _prompt?: Deferred<void>;
-    public static get prompt(): Promise<void> | undefined {
-        return JupyterPasswordConnect._prompt?.promise;
-    }
     public getPasswordConnectionInfo(options: {
         url: string;
         isTokenEmpty: boolean;
@@ -61,7 +49,6 @@ export class JupyterPasswordConnect {
         validationErrorMessage?: string;
         disposables?: IDisposable[];
     }): Promise<IJupyterPasswordConnectInfo> {
-        JupyterPasswordConnect._prompt = undefined;
         if (!options.url || options.url.length < 1) {
             throw new Error('Invalid URL');
         }
@@ -74,8 +61,7 @@ export class JupyterPasswordConnect {
         // See if we already have this data. Don't need to ask for a password more than once. (This can happen in remote when listing kernels)
         let result = this.savedConnectInfo.get(options.handle);
         if (!result) {
-            const deferred = (JupyterPasswordConnect._prompt = createDeferred());
-            result = this.getNonCachedPasswordConnectionInfo({
+            result = this.getJupyterConnectionInfo({
                 url: newUrl,
                 isTokenEmpty: options.isTokenEmpty,
                 displayName: options.displayName,
@@ -93,10 +79,6 @@ export class JupyterPasswordConnect {
             result.catch(() => this.savedConnectInfo.delete(options.handle));
             result
                 .finally(() => {
-                    deferred.resolve();
-                    if (JupyterPasswordConnect._prompt === deferred) {
-                        JupyterPasswordConnect._prompt = undefined;
-                    }
                     if (disposeOnDone) {
                         disposeAllDisposables(disposables);
                     }
@@ -106,200 +88,6 @@ export class JupyterPasswordConnect {
         }
 
         return result;
-    }
-
-    private async getNonCachedPasswordConnectionInfo(options: {
-        url: string;
-        isTokenEmpty: boolean;
-        displayName?: string;
-        validationErrorMessage?: string;
-        disposables: IDisposable[];
-    }): Promise<IJupyterPasswordConnectInfo> {
-        // If jupyter hub, go down a special path of asking jupyter hub for a token
-        if (await this.isJupyterHub(options.url)) {
-            return this.getJupyterHubConnectionInfo(options.url, options.validationErrorMessage);
-        } else {
-            return this.getJupyterConnectionInfo(options);
-        }
-    }
-
-    private async getJupyterHubConnectionInfo(
-        uri: string,
-        validationErrorMessage?: string
-    ): Promise<IJupyterPasswordConnectInfo> {
-        try {
-            // First ask for the user name and password
-            const userNameAndPassword = await this.getUserNameAndPassword(validationErrorMessage);
-            if (userNameAndPassword.username || userNameAndPassword.password) {
-                // Try the login method. It should work and doesn't require a token to be generated.
-                let result = await this.getJupyterHubConnectionInfoFromLogin(
-                    uri,
-                    userNameAndPassword.username,
-                    userNameAndPassword.password
-                );
-
-                // If login method fails, try generating a token
-                if (result) {
-                    const failed = Object.keys(result || {}).length === 0;
-                    const info = failed ? 'emptyResponseFromLogin' : 'gotResponseFromLogin';
-                    sendTelemetryEvent(Telemetry.CheckPasswordJupyterHub, undefined, {
-                        failed,
-                        info
-                    });
-                } else {
-                    sendTelemetryEvent(Telemetry.CheckPasswordJupyterHub, undefined, {
-                        failed: true,
-                        info: 'emptyResponseFromLogin'
-                    });
-                    result = await this.getJupyterHubConnectionInfoFromApi(
-                        uri,
-                        userNameAndPassword.username,
-                        userNameAndPassword.password
-                    );
-                    const failed = Object.keys(result || {}).length === 0;
-                    const info = failed ? 'emptyResponseFromApi' : 'gotResponseFromApi';
-                    sendTelemetryEvent(Telemetry.CheckPasswordJupyterHub, undefined, {
-                        failed,
-                        info
-                    });
-                }
-
-                return result;
-            }
-            sendTelemetryEvent(Telemetry.CheckPasswordJupyterHub, undefined, {
-                failed: false,
-                info: 'passwordNotRequired'
-            });
-            return {
-                requiresPassword: false
-            };
-        } catch (ex) {
-            sendTelemetryEvent(Telemetry.CheckPasswordJupyterHub, undefined, { failed: true, info: 'failure' });
-            throw ex;
-        }
-    }
-
-    private async getJupyterHubConnectionInfoFromLogin(
-        uri: string,
-        username: string,
-        password: string
-    ): Promise<IJupyterPasswordConnectInfo | undefined> {
-        // We're using jupyter hub. Get the base url
-        const url = new URL(uri);
-        const baseUrl = `${url.protocol}//${url.host}`;
-
-        const postParams = new URLSearchParams();
-        postParams.append('username', username || '');
-        postParams.append('password', password || '');
-
-        let response = await this.makeRequest(`${baseUrl}/hub/login?next=`, {
-            method: 'POST',
-            headers: {
-                Connection: 'keep-alive',
-                Referer: `${baseUrl}/hub/login`,
-                'content-type': 'application/x-www-form-urlencoded;charset=UTF-8'
-            },
-            body: postParams.toString(),
-            redirect: 'manual'
-        });
-
-        // The cookies from that response should be used to make the next set of requests
-        if (response && response.status === 302) {
-            const cookies = this.getCookies(response);
-            const cookieString = [...cookies.entries()].reduce((p, c) => `${p};${c[0]}=${c[1]}`, '');
-            // See this API for creating a token
-            // https://jupyterhub.readthedocs.io/en/stable/_static/rest-api/index.html#operation--users--name--tokens-post
-            response = await this.makeRequest(`${baseUrl}/hub/api/users/${username}/tokens`, {
-                method: 'POST',
-                headers: {
-                    Connection: 'keep-alive',
-                    Cookie: cookieString,
-                    Referer: `${baseUrl}/hub/login`
-                }
-            });
-
-            // That should give us a new token. For now server name is hard coded. Not sure
-            // how to fetch it other than in the info for a default token
-            if (response.ok && response.status === 200) {
-                const body = await response.json();
-                if (body && body.token && body.id) {
-                    // Response should have the token to use for this user.
-
-                    // Make sure the server is running for this user. Don't need
-                    // to check response as it will fail if already running.
-                    // https://jupyterhub.readthedocs.io/en/stable/_static/rest-api/index.html#operation--users--name--server-post
-                    await this.makeRequest(`${baseUrl}/hub/api/users/${username}/server`, {
-                        method: 'POST',
-                        headers: {
-                            Connection: 'keep-alive',
-                            Cookie: cookieString,
-                            Referer: `${baseUrl}/hub/login`
-                        }
-                    });
-
-                    // This token was generated for this request. We should clean it up when
-                    // the user closes VS code
-                    this.asyncDisposableRegistry.push({
-                        dispose: async () => {
-                            this.makeRequest(`${baseUrl}/hub/api/users/${username}/tokens/${body.id}`, {
-                                method: 'DELETE',
-                                headers: {
-                                    Connection: 'keep-alive',
-                                    Cookie: cookieString,
-                                    Referer: `${baseUrl}/hub/login`
-                                }
-                            }).catch(noop); // Don't wait for this during shutdown. Just make the request
-                        }
-                    });
-
-                    return {
-                        requestHeaders: {},
-                        remappedBaseUrl: `${baseUrl}/user/${username}`,
-                        remappedToken: body.token,
-                        requiresPassword: true
-                    };
-                }
-            }
-        }
-    }
-
-    private async getJupyterHubConnectionInfoFromApi(
-        uri: string,
-        username: string,
-        password: string
-    ): Promise<IJupyterPasswordConnectInfo> {
-        // We're using jupyter hub. Get the base url
-        const url = new URL(uri);
-        const baseUrl = `${url.protocol}//${url.host}`;
-        // Use these in a post request to get the token to use
-        const response = await this.makeRequest(
-            `${baseUrl}/hub/api/authorizations/token`, // This seems to be deprecated, but it works. It requests a new token
-            {
-                method: 'POST',
-                headers: {
-                    Connection: 'keep-alive',
-                    'content-type': 'application/json;charset=UTF-8'
-                },
-                body: `{ "username": "${username || ''}", "password": "${password || ''}"  }`,
-                redirect: 'manual'
-            }
-        );
-
-        if (response.ok && response.status === 200) {
-            const body = await response.json();
-            if (body && body.user && body.user.server && body.token) {
-                // Response should have the token to use for this user.
-                return {
-                    requestHeaders: {},
-                    remappedBaseUrl: `${baseUrl}${body.user.server}`,
-                    remappedToken: body.token,
-                    requiresPassword: true
-                };
-            }
-        }
-        return {
-            requiresPassword: false
-        };
     }
 
     /**
@@ -418,50 +206,6 @@ export class JupyterPasswordConnect {
         return options;
     }
 
-    private async getUserNameAndPassword(validationMessage?: string): Promise<{ username: string; password: string }> {
-        const multistep = this.multiStepFactory.create<{
-            username: string;
-            password: string;
-            validationMessage?: string;
-        }>();
-        const state = { username: '', password: '', validationMessage };
-        await multistep.run(this.getUserNameMultiStep.bind(this), state);
-        return state;
-    }
-
-    private async getUserNameMultiStep(
-        input: IMultiStepInput<{ username: string; password: string; validationErrorMessage?: string }>,
-        state: { username: string; password: string; validationMessage?: string }
-    ) {
-        state.username = await input.showInputBox({
-            title: DataScience.jupyterSelectUserAndPasswordTitle,
-            prompt: DataScience.jupyterSelectUserPrompt,
-            validate: this.validateUserNameOrPassword,
-            validationMessage: state.validationMessage,
-            value: ''
-        });
-        if (state.username) {
-            return this.getPasswordMultiStep.bind(this);
-        }
-    }
-
-    private async validateUserNameOrPassword(_value: string): Promise<string | undefined> {
-        return undefined;
-    }
-
-    private async getPasswordMultiStep(
-        input: IMultiStepInput<{ username: string; password: string }>,
-        state: { username: string; password: string }
-    ) {
-        state.password = await input.showInputBox({
-            title: DataScience.jupyterSelectUserAndPasswordTitle,
-            prompt: DataScience.jupyterSelectPasswordPrompt,
-            validate: this.validateUserNameOrPassword,
-            value: '',
-            password: true
-        });
-    }
-
     private async getXSRFToken(url: string, sessionCookie: string): Promise<string | undefined> {
         let xsrfCookie: string | undefined;
         let headers;
@@ -541,26 +285,6 @@ export class JupyterPasswordConnect {
             }
             throw e;
         }
-    }
-
-    private async isJupyterHub(url: string): Promise<boolean> {
-        // See this for the different REST endpoints:
-        // https://jupyterhub.readthedocs.io/en/stable/_static/rest-api/index.html
-
-        // If the URL has the /user/ option in it, it's likely this is jupyter hub
-        if (url.toLowerCase().includes('/user/')) {
-            return true;
-        }
-
-        // Otherwise request hub/api. This should return the json with the hub version
-        // if this is a hub url
-        const response = await this.makeRequest(`${url}hub/api`, {
-            method: 'get',
-            redirect: 'manual',
-            headers: { Connection: 'keep-alive' }
-        });
-
-        return response.status === 200;
     }
 
     /**
